+++ server/entities/Monster.js
@@ -1,26 +1,29 @@
 import { MONSTER_TYPES } from '../../shared/monsters.js';
+import { scaleMonsterDefinition } from '../../shared/plugins/combat/monsterScaling.js';
 import { DAMAGE_TYPE, createEmptyResistances } from '../../shared/plugins/combat/damageTypes.js';
 import { getBossPhase } from '../../shared/plugins/combat/bossPhases.js';
 
 let nextId = 1;
 
 export class Monster {
-  constructor({ type, x, y }) {
+  constructor({ type, x, y, scaleLevel = 1 }) {
     const def = MONSTER_TYPES[type];
+    const scaled = scaleMonsterDefinition(def, scaleLevel);
     this.id = `m${nextId++}`;
     this.type = type;
     this.label = def.label;
+    this.level = scaleLevel;
     this.x = x;
     this.y = y;
-    this.maxHp = def.hp;
-    this.hp = def.hp;
+    this.maxHp = scaled.hp;
+    this.hp = scaled.hp;
     this.speed = def.speed;
     this.baseSpeed = def.speed;
-    this.damage = def.damage;
+    this.damage = scaled.damage;
     this.aggroRange = def.aggroRange;
     this.attackRange = def.attackRange;
     this.color = def.color;
-    this.xpReward = def.xpReward ?? 10;
+    this.xpReward = scaled.xpReward;
     this.isBoss = !!def.isBoss;
     this.isElite = false;
     this.eliteModifier = null;
@@ -47,12 +50,13 @@ export class Monster {
       moving: this.moving,
       isBoss: this.isBoss,
       isElite: this.isElite,
+      level: this.level,
       eliteModifier: this.eliteModifier,
       bossPhase: this.isBoss ? getBossPhase(this) : undefined,
     };
   }
 }
 
-export function createMonster(type, x, y) {
-  return new Monster({ type, x, y });
+export function createMonster(type, x, y, scaleLevel = 1) {
+  return new Monster({ type, x, y, scaleLevel });
 }